+++ src/enums/cpu-usage-metric.enum.ts
@@ -1,0 +1,12 @@
+export enum CpuUsageMetric {
+  CPU_TIME = "process.cpu.time",
+}
+
+export enum CpuUsageAttribute {
+  MODE = "cpu.mode",
+}
+
+export enum CpuMode {
+  USER = "user",
+  SYSTEM = "system",
+}